+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/Body.hs
@@ -65,12 +65,16 @@ module Cardano.Ledger.Api.Tx.Body (
   ConwayEraTxBody,
   votingProceduresTxBodyL,
   proposalProceduresTxBodyL,
+
+  -- * Upgrade
+  binaryUpgradeTxBody,
+  upgradeTxBody,
 ) where
 
 import Cardano.Ledger.Address (Withdrawals (..))
 import Cardano.Ledger.Allegra.Core (AllegraEraTxBody (..))
 import Cardano.Ledger.Alonzo.TxBody (AlonzoEraTxBody (..))
-import Cardano.Ledger.Api.Era ()
+import Cardano.Ledger.Api.Era (EraApi (..))
 import Cardano.Ledger.Api.Scripts
 import Cardano.Ledger.Api.Tx.Out
 import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash)
@@ -78,7 +82,14 @@ import Cardano.Ledger.Babbage.TxBody (BabbageEraTxBody (..))
 import Cardano.Ledger.BaseTypes (SlotNo, StrictMaybe (..))
 import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Conway.TxBody (ConwayEraTxBody (..))
-import Cardano.Ledger.Core (EraTxBody (..), PParams, TxAuxDataHash (..), Value, txIdTxBody)
+import Cardano.Ledger.Core (
+  EraTxBody (..),
+  PParams,
+  TxAuxDataHash (..),
+  Value,
+  binaryUpgradeTxBody,
+  txIdTxBody,
+ )
 import Cardano.Ledger.Credential (Credential)
 import Cardano.Ledger.Keys (KeyHash (..), KeyRole (..))
 import Cardano.Ledger.Mary.Core (MaryEraTxBody (..))